Move GMSHInterface from FileIO to GeoLib/IO.
