+++ Applications/DataExplorer/VtkAct/CMakeLists.txt
@@ -12,7 +12,7 @@ include_directories(
     ${CMAKE_CURRENT_SOURCE_DIR}/../../BaseLib
     ${CMAKE_CURRENT_SOURCE_DIR}/../../GeoLib
     ${CMAKE_CURRENT_SOURCE_DIR}/../../FileIO
-    ${CMAKE_CURRENT_SOURCE_DIR}/../../FileIO/GmshIO
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../FileIO/Gmsh
     ${CMAKE_CURRENT_SOURCE_DIR}/../../MeshLib
     ${CMAKE_CURRENT_SOURCE_DIR}/../../MathLib
     ${CMAKE_CURRENT_SOURCE_DIR}/../VtkVis
