+++ Applications/DataExplorer/mainwindow.cpp
@@ -65,10 +65,10 @@
 #include "MeshLib/IO/Legacy/MeshIO.h"
 #include "MeshLib/IO/readMeshFromFile.h"
 #include "MeshLib/IO/GmshReader.h"
-#include "FileIO/GMSHInterface.h"
 #include "FileIO/TetGenInterface.h"
 #include "PetrelInterface.h"
 #include "XmlIO/Qt/XmlGspInterface.h"
+#include "GeoLib/IO/GMSHInterface.h"
 #include "GeoLib/IO/FEFLOW/FEFLOWGeoInterface.h"
 #include "GeoLib/IO/XmlIO/Qt/XmlGmlInterface.h"
 #include "GeoLib/IO/XmlIO/Qt/XmlStnInterface.h"
