+++ FileIO/CMakeLists.txt
@@ -1,7 +1,5 @@
 GET_SOURCE_FILES(SOURCES_FileIO)
 set(SOURCES ${SOURCES_FileIO})
-GET_SOURCE_FILES(SOURCES_GMSHIO GmshIO)
-set(SOURCES ${SOURCES} ${SOURCES_GMSHIO})
 
 # Create the library
 add_library(FileIO ${SOURCES})
