+++ GeoLib/CMakeLists.txt
@@ -18,6 +18,9 @@ if(QT4_FOUND)
     SET(SOURCES ${SOURCES} ${SOURCES_IO_FEFLOW})
 endif()
 
+GET_SOURCE_FILES(SOURCES_IO_GMSHIO IO/Gmsh)
+set(SOURCES ${SOURCES} ${SOURCES_IO_GMSHIO})
+
 # Create the library
 add_library(GeoLib STATIC ${SOURCES}
     ${CMAKE_CURRENT_SOURCE_DIR}/../ThirdParty/tetgen/predicates.cxx
