+++ GeoLib/IO/Gmsh/GMSHAdaptiveMeshDensity.h
@@ -1,8 +1,4 @@
 /**
- * \file
- * \author Thomas Fischer
- * \date   2012-03-05
- * \brief  Definition of the GMSHAdaptiveMeshDensity class.
  *
  * \copyright
  * Copyright (c) 2012-2016, OpenGeoSys Community (http://www.opengeosys.org)
