+++ GeoLib/IO/Gmsh/GMSHMeshDensityStrategy.h
@@ -1,8 +1,4 @@
 /**
- * \file
- * \author Thomas Fischer
- * \date   2010-03-05
- * \brief  Definition of the GMSHMeshDensityStrategy class.
  *
  * \copyright
  * Copyright (c) 2012-2016, OpenGeoSys Community (http://www.opengeosys.org)
