+++ GeoLib/IO/Gmsh/GMSHPolygonTree.h
@@ -1,8 +1,4 @@
 /**
- * \file
- * \author Thomas Fischer
- * \date   Mar 27 2012
- * \brief  Definition of the GMSHPolygonTree class.
  *
  * \copyright
  * Copyright (c) 2012-2016, OpenGeoSys Community (http://www.opengeosys.org)
